learnbydoing/requirements.txt
Django
django-rest-framework
psycopg2-binary
django-cors-headers
